Внесены изменения после ревью. Аргумент completion теперь передает Swift.Result<String, Error>, как по ТЗ
