+++ image_feed/Auth/OAuth2TokenStorage.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-class OAuth2TokenStorage {
+final class OAuth2TokenStorage {
     static let shared = OAuth2TokenStorage()
     
     private let userDefaults = UserDefaults.standard
