+++ image_feed/Constants.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import WebKit
 
 enum Constants {
     static let accessKey = "HyAv7Tq7aJRGyxa_-p_0yvJTH9rlpFNpLiEUkcBxjZ0"
@@ -13,5 +14,6 @@ enum Constants {
     static let redirectURI = "urn:ietf:wg:oauth:2.0:oob"
     static let accessScope = "public+read_user+write_likes"
     static let defaultBaseURL = URL(string: "https://api.unsplash.com")
+    static let unsplashAuthorizeURLString = "https://unsplash.com/oauth/authorize"
 }
