+++ image_feed/Models+Structs+Classes+Enums/OAuthError.swift
@@ -1,0 +1,13 @@
+//
+//  OAuthError.swift
+//  image_feed
+//
+//  Created by Aleksei Frolov on 06.04.2024.
+//
+
+import Foundation
+
+enum OAuthError: Error {
+        case invalidURL
+        case decodingError
+    }
