+++ image_feed/Models+Structs+Classes+Enums/OAuthTokenResponseBody.swift
@@ -1,0 +1,13 @@
+//
+//  OAuthTokenResponseBody.swift
+//  image_feed
+//
+//  Created by Aleksei Frolov on 06.04.2024.
+//
+
+import Foundation
+
+struct OAuthTokenResponseBody: Codable {
+    let accessToken: String
+    let tokenType: String
+}
